Implement routed model management UI
