+++ ui/package.json
@@ -14,6 +14,7 @@
     "lucide-react": "^1.8.0",
     "react": "^19.2.4",
     "react-dom": "^19.2.4",
+    "react-router-dom": "^7.14.1",
     "tailwind-merge": "^3.5.0"
   },
   "devDependencies": {
