+++ ui/src/components/ModelManagementDashboard.tsx
@@ -1,4 +1,3 @@
-import { startTransition, useEffect, useMemo, useState } from 'react'
 import {
   AlertTriangle,
   Download,
@@ -8,88 +7,22 @@ import {
   Trash2,
 } from 'lucide-react'
 
+import { useModelManagement, type ModelPanelItem } from '../context/ModelManagementContext'
 import { useI18n } from '../i18n/useI18n'
-import { fetchJson } from '../lib/api'
+import {
+  EMPTY_MODEL_VALUE,
+  formatModelBytes,
+  formatModelDate,
+  getModelStateTone,
+  getModelUiState,
+  getNoticeTone,
+} from '../lib/modelManagementUi'
 import { cn } from '../lib/utils'
-
-type ModelPanelStatus = 'downloaded' | 'not_downloaded' | 'unknown'
-type ModelLanguageScope = 'multilingual' | 'english'
-type ModelFamily = 'standard' | 'distil'
-type OperationJobStatus = 'queued' | 'running' | 'success' | 'error'
-type OperationType = 'download' | 'delete'
-type ModelCardUiState =
-  | 'downloaded'
-  | 'not_downloaded'
-  | 'unknown'
-  | 'downloading'
-  | 'download_error'
-  | 'deleting'
-  | 'delete_error'
-type NoticeKind = 'info' | 'success' | 'error'
-
-type ModelPanelItem = {
-  id: string
-  displayName: string
-  backendValue: string
-  hfRepoId: string
-  languageScope: ModelLanguageScope
-  family: ModelFamily
-  enabled: boolean
-  isDownloaded: boolean
-  cacheLocation: string | null
-  downloadedSizeBytes: number | null
-  lastModified: string | null
-  status: ModelPanelStatus
-}
-
-type ModelPanelResponse = {
-  catalog: ModelPanelItem[]
-  summary: {
-    downloadedCount: number
-    availableCount: number
-    totalDownloadedSizeBytes: number
-  }
-}
-
-type ModelOperationJobResponse = {
-  jobId: string
-  modelId: string
-  status: OperationJobStatus
-  progress: {
-    percent: number
-    label: string
-  }
-  message: string
-  error: string | null
-}
-
-type ActiveOperation = {
-  operationType: OperationType
-  jobId: string
-  modelId: string
-}
-
-type Notice = {
-  kind: NoticeKind
-  text: string
-}
-
-const EMPTY_SUMMARY: ModelPanelResponse['summary'] = {
-  downloadedCount: 0,
-  availableCount: 0,
-  totalDownloadedSizeBytes: 0,
-}
-
-const EMPTY_VALUE = '\u2014'
 
 const COPY = {
   en: {
-    headerEyebrow: 'Model Management',
-    headerTitle: 'Whisper model inventory',
-    headerDescription:
-      'Preload and remove Whisper models in the same local Hugging Face cache used by faster-whisper.',
     refreshing: 'Refreshing',
-    ready: 'Panel ready',
+    ready: 'Dashboard ready',
     operationBusy: 'Operation in progress',
     summaryDownloaded: 'Downloaded',
     summaryAvailable: 'Available',
@@ -98,7 +31,7 @@ const COPY = {
     loadingDescription: 'Requesting /api/models/panel and preparing the Whisper model inventory.',
     errorTitle: 'Model panel unavailable',
     emptyTitle: 'No models in catalog',
-    emptyDescription: 'The backend returned an empty catalog. The panel is waiting for registry data.',
+    emptyDescription: 'The backend returned an empty catalog. The dashboard is waiting for registry data.',
     hfCacheState: 'HF cache state',
     emptyErrorState: 'Empty / Error State',
     backendValue: 'backendValue',
@@ -112,14 +45,9 @@ const COPY = {
     percent: 'Percent',
     download: 'Download',
     downloadRunning: 'Downloading',
-    downloadReady: 'Model is available locally.',
-    downloadFailed: 'Download failed',
     delete: 'Delete',
     deleteRunning: 'Deleting',
-    deleteDone: 'Model removed from local cache.',
-    deleteFailed: 'Delete failed',
     deleteConfirm: (name: string) => `Delete the local model "${name}" from the Hugging Face cache?`,
-    queueMessage: 'The backend accepted the operation and started background processing.',
     anotherOperation: 'Another model operation is already running.',
     deleteDisabled: 'Delete is only available for downloaded models.',
     statusLabel: {
@@ -141,12 +69,8 @@ const COPY = {
     },
   },
   ru: {
-    headerEyebrow: 'Управление моделями',
-    headerTitle: 'Локальный каталог Whisper-моделей',
-    headerDescription:
-      'Предзагружайте и удаляйте Whisper-модели в том же локальном Hugging Face cache, который использует faster-whisper.',
     refreshing: 'Обновление',
-    ready: 'Панель готова',
+    ready: 'Dashboard готов',
     operationBusy: 'Идёт операция',
     summaryDownloaded: 'Загружено',
     summaryAvailable: 'Доступно',
@@ -155,7 +79,7 @@ const COPY = {
     loadingDescription: 'Запрашиваем /api/models/panel и собираем локальный каталог Whisper-моделей.',
     errorTitle: 'Панель моделей недоступна',
     emptyTitle: 'Каталог моделей пуст',
-    emptyDescription: 'Backend вернул пустой каталог. Панель ждёт данные из реестра моделей.',
+    emptyDescription: 'Backend вернул пустой каталог. Dashboard ждёт данные реестра моделей.',
     hfCacheState: 'Состояние HF cache',
     emptyErrorState: 'Пустое / ошибочное состояние',
     backendValue: 'backendValue',
@@ -169,14 +93,9 @@ const COPY = {
     percent: 'Процент',
     download: 'Скачать',
     downloadRunning: 'Загрузка',
-    downloadReady: 'Модель доступна локально.',
-    downloadFailed: 'Загрузка завершилась ошибкой',
     delete: 'Удалить',
     deleteRunning: 'Удаление',
-    deleteDone: 'Модель удалена из локального cache.',
-    deleteFailed: 'Удаление завершилось ошибкой',
     deleteConfirm: (name: string) => `Удалить локальную модель "${name}" из Hugging Face cache?`,
-    queueMessage: 'Backend принял задачу и запустил фоновую операцию.',
     anotherOperation: 'Сейчас уже выполняется другая операция с моделями.',
     deleteDisabled: 'Удаление доступно только для скачанных моделей.',
     statusLabel: {
@@ -198,116 +117,6 @@ const COPY = {
     },
   },
 } as const
-
-function formatBytes(value: number | null) {
-  if (value == null) {
-    return EMPTY_VALUE
-  }
-  if (value === 0) {
-    return '0 B'
-  }
-
-  const units = ['B', 'KB', 'MB', 'GB']
-  const exponent = Math.min(Math.floor(Math.log(value) / Math.log(1024)), units.length - 1)
-  const size = value / 1024 ** exponent
-
-  return `${size.toFixed(size >= 10 || exponent === 0 ? 0 : 1)} ${units[exponent]}`
-}
-
-function formatDate(value: string | null) {
-  if (!value) {
-    return EMPTY_VALUE
-  }
-
-  const parsed = new Date(value)
-  if (Number.isNaN(parsed.getTime())) {
-    return EMPTY_VALUE
-  }
-
-  return new Intl.DateTimeFormat(undefined, {
-    dateStyle: 'medium',
-    timeStyle: 'short',
-  }).format(parsed)
-}
-
-function getUiState(
-  item: ModelPanelItem,
-  downloadJob?: ModelOperationJobResponse,
-  deleteJob?: ModelOperationJobResponse,
-): ModelCardUiState {
-  if (deleteJob?.status === 'queued' || deleteJob?.status === 'running') {
-    return 'deleting'
-  }
-  if (deleteJob?.status === 'error') {
-    return 'delete_error'
-  }
-  if (downloadJob?.status === 'queued' || downloadJob?.status === 'running') {
-    return 'downloading'
-  }
-  if (downloadJob?.status === 'error') {
-    return 'download_error'
-  }
-
-  switch (item.status) {
-    case 'downloaded':
-      return 'downloaded'
-    case 'unknown':
-      return 'unknown'
-    default:
-      return 'not_downloaded'
-  }
-}
-
-function getStateTone(state: ModelCardUiState) {
-  switch (state) {
-    case 'downloaded':
-      return {
-        card: 'border-emerald-200 bg-emerald-50/50',
-        badge: 'border-emerald-200 bg-emerald-50 text-emerald-800',
-      }
-    case 'downloading':
-      return {
-        card: 'border-sky-200 bg-sky-50/55',
-        badge: 'border-sky-200 bg-sky-50 text-sky-800',
-      }
-    case 'download_error':
-      return {
-        card: 'border-rose-200 bg-rose-50/55',
-        badge: 'border-rose-200 bg-rose-50 text-rose-800',
-      }
-    case 'deleting':
-      return {
-        card: 'border-orange-200 bg-orange-50/60',
-        badge: 'border-orange-200 bg-orange-50 text-orange-800',
-      }
-    case 'delete_error':
-      return {
-        card: 'border-rose-200 bg-rose-50/55',
-        badge: 'border-rose-200 bg-rose-50 text-rose-800',
-      }
-    case 'unknown':
-      return {
-        card: 'border-amber-200 bg-amber-50/55',
-        badge: 'border-amber-200 bg-amber-50 text-amber-800',
-      }
-    default:
-      return {
-        card: 'border-stone-200 bg-stone-50/70',
-        badge: 'border-stone-200 bg-stone-100 text-stone-700',
-      }
-  }
-}
-
-function getNoticeTone(kind: NoticeKind) {
-  switch (kind) {
-    case 'success':
-      return 'border-emerald-200 bg-emerald-50 text-emerald-800'
-    case 'error':
-      return 'border-rose-200 bg-rose-50 text-rose-800'
-    default:
-      return 'border-sky-200 bg-sky-50 text-sky-800'
-  }
-}
 
 function SummaryCard(props: { label: string; value: string | number }) {
   return (
@@ -336,129 +145,37 @@ function InfoField(props: { label: string; value: string; title?: string; valueC
   )
 }
 
-export function ModelManagementPanel() {
+export function ModelManagementDashboard() {
   const { locale } = useI18n()
   const copy = locale === 'ru' ? COPY.ru : COPY.en
-  const [panelData, setPanelData] = useState<ModelPanelResponse | null>(null)
-  const [isLoading, setIsLoading] = useState(true)
-  const [error, setError] = useState<string | null>(null)
-  const [activeOperation, setActiveOperation] = useState<ActiveOperation | null>(null)
-  const [downloadJobs, setDownloadJobs] = useState<Record<string, ModelOperationJobResponse>>({})
-  const [deleteJobs, setDeleteJobs] = useState<Record<string, ModelOperationJobResponse>>({})
-  const [modelNotices, setModelNotices] = useState<Record<string, Notice>>({})
-  const [globalNotice, setGlobalNotice] = useState<Notice | null>(null)
-
-  async function loadPanel(options?: { keepLoadingState?: boolean }) {
-    if (!options?.keepLoadingState) {
-      setIsLoading(true)
-    }
-
-    try {
-      const payload = await fetchJson<ModelPanelResponse>('/api/models/panel')
-      startTransition(() => {
-        setPanelData(payload)
-        setError(null)
-      })
-    } catch (loadError) {
-      setError(loadError instanceof Error ? loadError.message : 'Unable to load model catalog.')
-    } finally {
-      setIsLoading(false)
-    }
-  }
-
-  useEffect(() => {
-    void loadPanel()
-  }, [])
-
-  useEffect(() => {
-    if (!activeOperation) {
-      return
-    }
-
-    let active = true
-    const timer = window.setInterval(async () => {
-      try {
-        const endpoint =
-          activeOperation.operationType === 'download'
-            ? `/api/models/download/${activeOperation.jobId}`
-            : `/api/models/delete/${activeOperation.jobId}`
-        const job = await fetchJson<ModelOperationJobResponse>(endpoint)
-        if (!active) {
-          return
+  const modelStatusLabels =
+    locale === 'ru'
+      ? {
+          downloaded: 'загружена',
+          not_downloaded: 'не загружена',
+          unknown: 'неизвестно',
+          downloading: 'загружается',
+          download_error: 'ошибка загрузки',
+          deleting: 'удаляется',
+          delete_error: 'ошибка удаления',
         }
-
-        startTransition(() => {
-          if (activeOperation.operationType === 'download') {
-            setDownloadJobs((current) => ({
-              ...current,
-              [job.modelId]: job,
-            }))
-          } else {
-            setDeleteJobs((current) => ({
-              ...current,
-              [job.modelId]: job,
-            }))
-          }
-        })
-
-        if (job.status === 'success') {
-          const successMessage =
-            activeOperation.operationType === 'download'
-              ? job.message || copy.downloadReady
-              : job.message || copy.deleteDone
-          setActiveOperation(null)
-          setModelNotices((current) => ({
-            ...current,
-            [job.modelId]: { kind: 'success', text: successMessage },
-          }))
-          setGlobalNotice({ kind: 'success', text: successMessage })
-          await loadPanel({ keepLoadingState: true })
-        } else if (job.status === 'error') {
-          const failureMessage =
-            job.error ||
-            job.message ||
-            (activeOperation.operationType === 'download' ? copy.downloadFailed : copy.deleteFailed)
-          setActiveOperation(null)
-          setModelNotices((current) => ({
-            ...current,
-            [job.modelId]: { kind: 'error', text: failureMessage },
-          }))
-        }
-      } catch (pollError) {
-        if (!active) {
-          return
-        }
-
-        setActiveOperation(null)
-        setGlobalNotice({
-          kind: 'error',
-          text:
-            pollError instanceof Error
-              ? pollError.message
-              : activeOperation.operationType === 'download'
-                ? copy.downloadFailed
-                : copy.deleteFailed,
-        })
-      }
-    }, 1000)
-
-    return () => {
-      active = false
-      window.clearInterval(timer)
-    }
-  }, [
+      : copy.statusLabel
+  const {
+    catalog,
+    summary,
+    isLoading,
+    error,
     activeOperation,
-    copy.deleteDone,
-    copy.deleteFailed,
-    copy.downloadFailed,
-    copy.downloadReady,
-  ])
+    downloadJobs,
+    deleteJobs,
+    modelNotices,
+    globalNotice,
+    startOperation,
+  } = useModelManagement()
 
-  const catalog = panelData?.catalog ?? []
-  const summary = panelData?.summary ?? EMPTY_SUMMARY
   const shouldShowModelList = catalog.length > 0
 
-  const stateMessage = useMemo(() => {
+  const stateMessage = (() => {
     if (isLoading) {
       return {
         title: copy.loadingTitle,
@@ -477,88 +194,7 @@ export function ModelManagementPanel() {
       title: copy.emptyTitle,
       description: copy.emptyDescription,
     }
-  }, [copy.emptyDescription, copy.emptyTitle, copy.errorTitle, copy.loadingDescription, copy.loadingTitle, error, isLoading])
-
-  function clearModelNotice(modelId: string) {
-    setModelNotices((current) => {
-      const next = { ...current }
-      delete next[modelId]
-      return next
-    })
-  }
-
-  async function startOperation(modelId: string, operationType: OperationType) {
-    clearModelNotice(modelId)
-    setGlobalNotice(null)
-
-    try {
-      const job = await fetchJson<ModelOperationJobResponse>(`/api/models/${operationType}`, {
-        method: 'POST',
-        headers: {
-          'Content-Type': 'application/json',
-        },
-        body: JSON.stringify({ modelId }),
-      })
-
-      startTransition(() => {
-        if (operationType === 'download') {
-          setDownloadJobs((current) => ({
-            ...current,
-            [job.modelId]: job,
-          }))
-        } else {
-          setDeleteJobs((current) => ({
-            ...current,
-            [job.modelId]: job,
-          }))
-        }
-      })
-
-      if (job.status === 'queued' || job.status === 'running') {
-        setActiveOperation({
-          operationType,
-          jobId: job.jobId,
-          modelId: job.modelId,
-        })
-        setModelNotices((current) => ({
-          ...current,
-          [job.modelId]: { kind: 'info', text: job.message || copy.queueMessage },
-        }))
-        return
-      }
-
-      if (job.status === 'success') {
-        const successMessage =
-          operationType === 'download' ? job.message || copy.downloadReady : job.message || copy.deleteDone
-        setModelNotices((current) => ({
-          ...current,
-          [job.modelId]: { kind: 'success', text: successMessage },
-        }))
-        setGlobalNotice({ kind: 'success', text: successMessage })
-        await loadPanel({ keepLoadingState: true })
-        return
-      }
-
-      const failureMessage =
-        job.error || job.message || (operationType === 'download' ? copy.downloadFailed : copy.deleteFailed)
-      setModelNotices((current) => ({
-        ...current,
-        [job.modelId]: { kind: 'error', text: failureMessage },
-      }))
-    } catch (operationError) {
-      const message =
-        operationError instanceof Error
-          ? operationError.message
-          : operationType === 'download'
-            ? copy.downloadFailed
-            : copy.deleteFailed
-      setGlobalNotice({ kind: 'error', text: message })
-      setModelNotices((current) => ({
-        ...current,
-        [modelId]: { kind: 'error', text: message },
-      }))
-    }
-  }
+  })()
 
   async function handleDownload(modelId: string) {
     await startOperation(modelId, 'download')
@@ -579,10 +215,12 @@ export function ModelManagementPanel() {
           <div className="flex flex-wrap items-start justify-between gap-3">
             <div className="max-w-3xl">
               <div className="text-xs font-semibold uppercase tracking-[0.2em] text-stone-500">
-                {copy.headerEyebrow}
+                API: /api/models/panel
               </div>
-              <h2 className="mt-2 text-2xl font-semibold text-stone-900">{copy.headerTitle}</h2>
-              <p className="mt-2 text-sm text-stone-500">{copy.headerDescription}</p>
+              <h2 className="mt-2 text-2xl font-semibold text-stone-900">ModelManagementDashboard</h2>
+              <p className="mt-2 text-sm text-stone-500">
+                Detailed inventory and local cache actions are rendered from the existing backend panel and operation endpoints.
+              </p>
             </div>
             <div className="rounded-full border border-stone-200 bg-stone-50 px-3 py-1 text-xs font-semibold uppercase tracking-[0.18em] text-stone-600">
               {isLoading ? copy.refreshing : activeOperation ? copy.operationBusy : copy.ready}
@@ -592,7 +230,7 @@ export function ModelManagementPanel() {
           <div className="grid gap-3 md:grid-cols-3">
             <SummaryCard label={copy.summaryDownloaded} value={summary.downloadedCount} />
             <SummaryCard label={copy.summaryAvailable} value={summary.availableCount} />
-            <SummaryCard label={copy.summaryTotalSize} value={formatBytes(summary.totalDownloadedSizeBytes)} />
+            <SummaryCard label={copy.summaryTotalSize} value={formatModelBytes(summary.totalDownloadedSizeBytes)} />
           </div>
 
           {globalNotice ? (
@@ -607,8 +245,8 @@ export function ModelManagementPanel() {
             {catalog.map((item) => {
               const downloadJob = downloadJobs[item.id]
               const deleteJob = deleteJobs[item.id]
-              const uiState = getUiState(item, downloadJob, deleteJob)
-              const tone = getStateTone(uiState)
+              const uiState = getModelUiState(item, downloadJob, deleteJob)
+              const tone = getModelStateTone(uiState)
               const notice = modelNotices[item.id]
               const isOperationActiveForThisModel =
                 activeOperation != null && activeOperation.modelId === item.id
@@ -638,27 +276,27 @@ export function ModelManagementPanel() {
                         tone.badge,
                       )}
                     >
-                      {copy.statusLabel[uiState]}
+                      {modelStatusLabels[uiState]}
                     </div>
                   </div>
 
                   <div className="mt-4 grid gap-3 md:grid-cols-2 xl:grid-cols-4">
                     <InfoField label={copy.family} value={copy.familyLabel[item.family]} />
                     <InfoField label={copy.languageScope} value={copy.languageLabel[item.languageScope]} />
-                    <InfoField label={copy.size} value={formatBytes(item.downloadedSizeBytes)} />
-                    <InfoField label={copy.lastModified} value={formatDate(item.lastModified)} />
+                    <InfoField label={copy.size} value={formatModelBytes(item.downloadedSizeBytes)} />
+                    <InfoField label={copy.lastModified} value={formatModelDate(item.lastModified)} />
                   </div>
 
                   <div className="mt-3 grid gap-3 xl:grid-cols-[minmax(0,1fr)_280px]">
                     <InfoField
                       label={copy.hfRepo}
-                      value={item.hfRepoId || EMPTY_VALUE}
+                      value={item.hfRepoId || EMPTY_MODEL_VALUE}
                       title={item.hfRepoId || undefined}
                       valueClassName="truncate"
                     />
                     <InfoField
                       label={copy.localPath}
-                      value={item.cacheLocation ?? EMPTY_VALUE}
+                      value={item.cacheLocation ?? EMPTY_MODEL_VALUE}
                       title={item.cacheLocation ?? undefined}
                       valueClassName="truncate"
                     />
