+++ ui/src/main.tsx
@@ -1,5 +1,6 @@
 import { StrictMode } from 'react'
 import { createRoot } from 'react-dom/client'
+import { BrowserRouter } from 'react-router-dom'
 import './index.css'
 import App from './App.tsx'
 import { I18nProvider } from './i18n/I18nProvider.tsx'
@@ -7,7 +8,9 @@ import { I18nProvider } from './i18n/I18nProvider.tsx'
 createRoot(document.getElementById('root')!).render(
   <StrictMode>
     <I18nProvider>
-      <App />
+      <BrowserRouter>
+        <App />
+      </BrowserRouter>
     </I18nProvider>
   </StrictMode>,
 )
